少许修改，增加文章摘要功能

+++ catsup.py
@@ -12,14 +12,25 @@ import tornado.template
 
 
 from tornado.options import define, options
-from utils import *
+from tornado.util import ObjectDict
+from utils import load_config, load_posts, get_infos
 
 define("port", default=8888, help="run on the given port", type=int)
 
 config = load_config()
-
+# posts = load_posts(config)
+# tags, archives = get_infos(posts)
 
 class BaseHandler(tornado.web.RequestHandler):
+
+    def render_string(self, template_name, **kwargs):
+        config.tags = self.settings['tags']
+        config.posts = self.settings['posts']
+        config.archives = self.settings['archives']
+        # access config directly
+        kwargs["config"] = config
+        return super(BaseHandler, self).render_string(template_name,**kwargs)
+
     def get_error_html(self, *args, **kwargs):
         return self.render_string('404.html')
 
@@ -34,6 +45,7 @@ class MainHandler(BaseHandler):
 
 class ArticleHandler(BaseHandler):
     def get(self, file_name):
+        posts = self.settings['posts']
         posts_num = len(posts)
         prev = next = None
         for i in range(posts_num):
@@ -55,12 +67,13 @@ class TagsHandler(BaseHandler):
 
 class TagHandler(BaseHandler):
     def get(self, tag_name):
+        tags = self.settings['tags']
         prev = next = None
-        for i, tag in enumerate(self.settings['tags']):
+        for i, tag in enumerate(tags):
             if tag[0] == tag_name:
                 i += 1
-                if i < len(self.settings['tags']):
-                    next = self.settings['tags'][i]
+                if i < len(tags):
+                    next = tags[i]
                 return self.render('tag.html', tag=tag,
                     prev=prev, next=next)
             prev = tag
@@ -75,12 +88,13 @@ class ArchivesHandler(BaseHandler):
 
 class ArchiveHandler(BaseHandler):
     def get(self, archive_name):
+        archives = self.settings['archives']
         prev = next = None
-        for i, archive in enumerate(self.settings['archives']):
+        for i, archive in enumerate(archives):
             if archive[0] == archive_name:
                 i += 1
-                if i < len(self.settings['archives']):
-                    next = self.settings['archives'][i]
+                if i < len(archives):
+                    next = archives[i]
                 return self.render('archive.html', archive=archive,
                     prev=prev, next=next)
             prev = archive
@@ -97,8 +111,8 @@ class FeedHandler(BaseHandler):
         self.set_header("Content-Type", "application/atom+xml")
         loader = tornado.template.Loader(config['common_template_path'],
             autoescape=None)
-        p = loader.load("feed.xml").generate(posts=self.settings['posts'],
-            handler=handler)
+        p = loader.load("feed.xml").generate(posts=self.settings['posts'], 
+            config=config)
         self.write(p)
 
 
@@ -108,7 +122,7 @@ class SitemapHandler(BaseHandler):
         loader = tornado.template.Loader(config['common_template_path'],
             autoescape=None)
         p = loader.load("sitemap.txt").generate(posts=self.settings['posts'],
-            handler=handler, tags=self.settings['tags'],
+            config=config, tags=self.settings['tags'],
             archives=self.settings['archives'])
         self.write(p)
 
@@ -136,12 +150,11 @@ class ErrorHandler(BaseHandler):
         raise tornado.web.HTTPError(404)
 
 
-posts = load_posts(config)
-tags, archives = get_infos(posts)
-
-
 def write(file_name, page):
-    file_path = os.path.join(config['deploy_path'], file_name)
+    if not file_name.startswith(config['deploy_path']):
+        file_path = os.path.join(config['deploy_path'], file_name)
+    else:
+        file_path = file_name
     open(file_path, 'w').write(page)
 
 
@@ -149,9 +162,9 @@ def deploy():
     posts = load_posts(config)
     tags, archives = get_infos(posts)
     posts_num = len(posts)
+    config['posts'] = posts
     config['tags'] = tags
     config['archives'] = archives
-    handler = FakeHandler(config)
 
     if os.path.exists(config.deploy_path):
         shutil.rmtree(config.deploy_path)
@@ -162,28 +175,32 @@ def deploy():
         autoescape=None)
 
     print('Generating sitemap')
-    page = loader.load("sitemap.txt").generate(posts=posts, handler=handler,
-        tags=tags, archives=archives)
+    page = loader.load("sitemap.txt").generate(posts=posts, tags=tags, 
+        archives=archives, config=config)
     write('sitemap.txt', page)
 
     print('Generating atom')
-    page = loader.load("feed.xml").generate(posts=posts, handler=handler)
+    page = loader.load("feed.xml").generate(posts=posts, config=config)
     write('feed.xml', page)
 
     loader = tornado.template.Loader(config['template_path'],
         autoescape=None)
 
     print('Start generating index pages')
+    page_path = os.path.join(config.deploy_path, 'page')
+    if os.path.exists(page_path):
+        shutil.rmtree(page_path)
+    os.makedirs(page_path)
     generator = loader.load("index.html")
     p = 0
     while posts_num > p * config['post_per_page']:
         p += 1
         print('Start generating page %s' % p)
-        page = generator.generate(posts=posts, handler=handler,
-            p=p)
-        write('page_%s.html' % p, page)
+        page = generator.generate(posts=posts, p=p, config=config)
+        pager_file = os.path.join(page_path, "%s.html" % p)
+        write(pager_file, page)
 
-    index_1 = os.path.join(config.deploy_path, 'page_1.html')
+    index_1 = os.path.join(page_path, '1.html')
     index = os.path.join(config.deploy_path, 'index.html')
     os.rename(index_1, index)
 
@@ -195,12 +212,16 @@ def deploy():
     next = len(posts) and posts.pop() or None
     while post:
         print('Generating %s' % post['file_name'])
-        page = generator.generate(post=post, handler=handler,
-            prev=prev, next=next)
+        page = generator.generate(post=post, prev=prev, 
+            next=next, config=config)
         write('%s.html' % post['file_name'], page)
         prev, post, next = post, next, len(posts) and posts.pop() or None
 
     print('Start generating tag pages')
+    tag_path = os.path.join(config.deploy_path, 'tag')
+    if os.path.exists(tag_path):
+        shutil.rmtree(tag_path)
+    os.makedirs(tag_path)
     generator = loader.load("tag.html")
     prev = None
     for i, tag in enumerate(tags):
@@ -208,11 +229,16 @@ def deploy():
         i += 1
         next = i < len(tags) and tags[i] or None
         page = generator.generate(tag=tag, prev=prev,
-            next=next, handler=handler)
-        write('tag_%s.html' % tag[0], page)
+            next=next, config=config)
+        tag_file = os.path.join(tag_path, "%s.html" % tag[0].lower())
+        write(tag_file, page)
         prev = tag
 
     print('Start generating archive pages')
+    archive_path = os.path.join(config.deploy_path, 'archive')
+    if os.path.exists(archive_path):
+        shutil.rmtree(archive_path)
+    os.makedirs(archive_path)
     generator = loader.load("archive.html")
     prev = None
     for i, archive in enumerate(archives):
@@ -220,13 +246,14 @@ def deploy():
         i += 1
         next = i < len(archives) and archives[i] or None
         page = generator.generate(archive=archive, prev=prev,
-            next=next, handler=handler)
-        write('archive_%s.html' % archive[0], page)
+            next=next, config=config)
+        archive_file = os.path.join(archive_path, "%s.html" % archive[0])
+        write(archive_file, page)
         prev = archive
 
     print('Start generating other pages')
     for p in ('404', 'tags', 'archives', 'links'):
-        page = loader.load("%s.html" % p).generate(handler=handler)
+        page = loader.load("%s.html" % p).generate(config=config)
         write('%s.html' % p, page)
 
     print('Copying static files.')
@@ -235,8 +262,18 @@ def deploy():
         shutil.rmtree(deploy_static_dir)
     shutil.copytree(config['static_path'], deploy_static_dir)
     os.chdir(config.deploy_path)
-    os.system('cp static/favicon.ico ./')
-    os.system('cp static/robots.txt ./')
+    # Favicon, use favicon.ico in _posts directory default or fallback to the one in static directory
+    favicon_file = os.path.join(config.posts_path, 'favicon.ico')
+    if os.path.exists(favicon_file):
+        os.system('cp %s ./' % favicon_file)
+    else:
+        os.system('cp static/favicon.ico ./')
+    # Robots.txt, use robots.txt in _posts directory default or fallback to the one in static directory
+    robots_file = os.path.join(config.posts_path, 'robots.txt')
+    if os.path.exists(robots_file):
+        os.system('cp %s ./' % robots_file)
+    else:
+        os.system('cp static/robots.txt ./')
 
     print('Done.')
 
@@ -257,13 +294,15 @@ if __name__ == '__main__':
     cmd = args[1]
     del args[1]
     if cmd == 'server':
+        posts = load_posts(config)
+        tags, archives = get_infos(posts)
         application = tornado.web.Application([
             (r'/', MainHandler),
-            (r'/page_(.*?).html', MainHandler),
+            (r'/page/(.*?).html', MainHandler),
             (r'/archives.html', ArchivesHandler),
-            (r'/archive_(.*?).html', ArchiveHandler),
+            (r'/archive/(.*?).html', ArchiveHandler),
             (r'/tags.html', TagsHandler),
-            (r'/tag_(.*?).html', TagHandler),
+            (r'/tag/(.*?).html', TagHandler),
             (r'/links.html', LinksHandler),
             (r'/feed.xml', FeedHandler),
             (r'/sitemap.txt', SitemapHandler),
@@ -284,6 +323,10 @@ if __name__ == '__main__':
 
     if 'application' in locals():
         tornado.options.parse_command_line(args)
+        if cmd == 'server':
+            print('Starting server at port %s' % options.port)
+        elif cmd == 'webhook':
+            print('Starting webhook at port %s' % options.port)
         http_server = tornado.httpserver.HTTPServer(application, xheaders=True)
         http_server.listen(options.port)
         tornado.ioloop.IOLoop.instance().start()
